+++ src/kitcube-devices/daqasciidevice.cpp
@@ -97,7 +97,7 @@ void DAQAsciiDevice::readData(const char *dir, const char *filename){
 	FILE *fd_data_file;
 	int j, k;
 	//char *sensorString;
-	float* local_sensorValue;
+	double *local_sensorValue;
 	//int err;
 	//std::string timeString;
 	//std::string dateString;
@@ -164,9 +164,9 @@ void DAQAsciiDevice::readData(const char *dir, const char *filename){
 	
 	// Allocate memory for sensor values
 	if (profile_length != 0) {
-		local_sensorValue = new float [nSensors * profile_length];
+		local_sensorValue = new double [nSensors * profile_length];
 	} else {
-		local_sensorValue = new float [nSensors];
+		local_sensorValue = new double [nSensors];
 	}
 	
 	if (debug >= 1)
@@ -234,7 +234,7 @@ void DAQAsciiDevice::readData(const char *dir, const char *filename){
 					}
 				} else {
 					for (j = 0; j < nSensors; j++) {
-						printf("%10.3f ", local_sensorValue[j]);
+						printf("%f ", local_sensorValue[j]);
 					}
 				}
 				printf("\n");
