+++ src/kitcube-devices/mast.h
@@ -65,7 +65,7 @@ public:
 	
 	void writeHeader();
 	
-	void parseData(char *line, struct timeval *l_tData, float * sensorValue);
+	void parseData(char *line, struct timeval *l_tData, double *sensorValue);
 	
 	/** Replace the time stamp of the data set by the current time */
 	void updateDataSet(unsigned char *buf);
